requirements.txt
discord.py
dpytest
pymongo